# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CineWeave API
  version: 0.1.0
  description: API contract seed for the Phase 1 CineWeave control plane.
servers:
  - url: http://localhost:8080
paths:
...
  /api/workflow-runs:
    get:
      operationId: listWorkflowRuns
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationHeader"
        - name: filter[projectId]
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Workflow runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunListEnvelope"
    post:
      operationId: createWorkflowRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRunRequest"
      responses:
        "201":
          description: Created workflow run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunEnvelope"
        "200":
          description: Idempotent replay of a previously created workflow run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunEnvelope"
        "422":
          $ref: "#/components/responses/Error"
  /api/workflow-runs/{workflowRunId}:
    get:
      operationId: getWorkflowRun
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkflowRunId"
      responses:
        "200":
          description: Workflow run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunEnvelope"
        "404":
          $ref: "#/components/responses/Error"
  /api/workflow-runs/{workflowRunId}/cancel:
    post:
      operationId: cancelWorkflowRun
      summary: Cancel a workflow run
      description: Sets a queued or running workflow to cancelling and requests Temporal cancellation. Terminal workflows return their current state.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkflowRunId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelWorkflowRunRequest"
      responses:
        "200":
          description: Workflow run after cancellation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunEnvelope"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/workflow-runs/{workflowRunId}/nodes:
    get:
      operationId: listWorkflowNodeRuns
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkflowRunId"
      responses:
        "200":
          description: Workflow node runs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowNodeRunListEnvelope"
        "404":
          $ref: "#/components/responses/Error"
...
  /api/artifacts:
    get:
      operationId: listArtifacts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrganizationHeader"
        - name: filter[projectId]
          in: query
          schema:
            type: string
            format: uuid
        - name: includePreviewUrl
          in: query
          schema:
            type: boolean
            default: false
        - name: previewExpiresSeconds
          in: query
          schema:
            type: integer
            default: 900
            maximum: 3600
      responses:
        "200":
          description: Project artifacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactListEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: service-token
  parameters:
...
    WorkflowRunId:
      name: workflowRunId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    OrganizationHeader:
      name: X-Organization-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 200
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Envelope:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
        data: {}
        meta: {}
...
    ErrorEnvelope:
      type: object
      required: [requestId, error]
      properties:
        requestId:
          type: string
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
          description: Includes PROVIDER_RATE_LIMITED, PROVIDER_CONCURRENCY_LIMITED, PROVIDER_DAILY_QUOTA_EXCEEDED, PROVIDER_MONTHLY_BUDGET_EXCEEDED, and PROVIDER_CIRCUIT_OPEN for Provider Gateway guard blocks.
        message:
          type: string
        details: {}
        retryable:
          type: boolean
...
    WorkflowRun:
      type: object
      required: [id, organizationId, projectId, temporalWorkflowId, status, input, output, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        temporalWorkflowId:
          type: string
        status:
          type: string
          enum: [queued, running, cancelling, succeeded, failed, cancelled]
        input:
          type: object
          additionalProperties: true
        output:
          description: Workflow output. text_to_storyboard returns TextToStoryboardOutput; video_production returns VideoProductionOutput.
          oneOf:
            - $ref: "#/components/schemas/TextToStoryboardOutput"
            - $ref: "#/components/schemas/VideoProductionOutput"
            - type: object
              additionalProperties: true
        errorCode:
          type: string
        errorMessage:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
    WorkflowNodeRun:
      type: object
      required: [id, organizationId, projectId, workflowRunId, nodeKey, nodeType, status, input, output, retryCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        nodeKey:
          type: string
        nodeType:
          type: string
          description: Workflow node type such as provider_text, image.generate, video.create_task, or media.compose.
        status:
          type: string
          enum: [pending, queued, running, succeeded, failed, cancelled, skipped, waiting_review]
        input:
          type: object
          additionalProperties: true
        output:
          description: Node progress or terminal output. Provider video nodes include providerAsyncTaskId, externalTaskId, and generated video artifact/media/storage fields after success.
          type: object
          additionalProperties: true
        retryCount:
          type: integer
        errorCode:
          type: string
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    TextToStoryboardOutput:
      type: object
      required: [storyboardArtifactId, providerCalls]
      properties:
        storyboardArtifactId:
          type: string
          format: uuid
        shots:
          type: array
          items:
            $ref: "#/components/schemas/StoryboardShot"
        imageArtifactId:
          type: string
          format: uuid
        imageMediaFileId:
          type: string
          format: uuid
        imageStorageKey:
          type: string
        providerCalls:
          type: object
          required: [storyboard]
          properties:
            storyboard:
              type: string
              format: uuid
            image:
              type: string
              format: uuid
    VideoProductionOutput:
      type: object
      required:
        - storyboardArtifactId
        - shots
        - providerCalls
      properties:
        storyboardArtifactId:
          type: string
          format: uuid
        shots:
          type: array
          items:
            $ref: "#/components/schemas/VideoProductionShotOutput"
        finalVideoArtifactId:
          type: string
          format: uuid
        finalVideoMediaFileId:
          type: string
          format: uuid
        finalVideoStorageKey:
          type: string
        timelineArtifactId:
          type: string
          format: uuid
        imageArtifactId:
          type: string
          format: uuid
        imageMediaFileId:
          type: string
          format: uuid
        imageStorageKey:
          type: string
        videoArtifactId:
          type: string
          format: uuid
        videoMediaFileId:
          type: string
          format: uuid
        videoStorageKey:
          type: string
        providerAsyncTaskId:
          type: string
          format: uuid
        externalTaskId:
          type: string
        providerCalls:
          type: object
          required: [storyboard]
          properties:
            storyboard:
              type: string
              format: uuid
            images:
              type: array
              items:
                type: string
                format: uuid
            videoCreates:
              type: array
              items:
                type: string
                format: uuid
            videoPolls:
              type: array
              items:
                type: string
                format: uuid
            image:
              type: string
              format: uuid
            videoCreate:
              type: string
              format: uuid
            videoPoll:
              type: string
              format: uuid
...
    Artifact:
      type: object
      required: [id, organizationId, type, metadata, createdAt]
      properties:
        id:
          type: string
          format: uuid
        organizationId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        workflowRunId:
          type: string
          format: uuid
        nodeRunId:
          type: string
          format: uuid
        type:
          type: string
          description: Artifact type. Workflow video production can emit storyboard_json, generated_image, generated_video, timeline_json, and final_video.
        storageKey:
          type: string
        mimeType:
          type: string
        contentHash:
          type: string
        promptHash:
          type: string
        modelId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        previewUrl:
          type: string
          format: uri
        previewExpiresAt:
          type: string
          format: date-time
...
    WorkflowRunEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/WorkflowRun"
    WorkflowRunListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/WorkflowRun"
    WorkflowNodeRunListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/WorkflowNodeRun"
...
    ArtifactListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/Artifact"
...
    CreateWorkflowRunRequest:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: string
          format: uuid
        workflowType:
          type: string
          enum: [video_production, text_to_storyboard, extract_novel_events, generate_adaptation_plan, adaptation_plan_to_script, source_to_script, parse_script_scenes, script_to_assets, script_to_storyboard, script_to_video, full_production, regenerate_canonical_asset_image, regenerate_derived_asset_image, regenerate_shot_image, regenerate_shot_video, regenerate_final_video, regenerate_script_scene, regenerate_scene_storyboard]
          default: video_production
          description: video_production keeps the prompt-only path unless input.scriptId is provided.
        prompt:
          type: string
        input:
          type: object
          additionalProperties: true
          description: Optional workflow-specific input. source_to_script, extract_novel_events, and generate_adaptation_plan require sourceId. adaptation_plan_to_script requires planId. video_production supports duration, aspectRatio, resolution, pollIntervalSeconds, maxPolls, maxShots, and skipCompose.
          properties:
            sourceId:
              type: string
              format: uuid
            planId:
              type: string
              format: uuid
            eventIds:
              type: array
              items:
                type: string
                format: uuid
            chapterIds:
              type: array
              items:
                type: string
                format: uuid
            force:
              type: boolean
            instruction:
              type: string
            title:
              type: string
            scriptId:
              type: string
              format: uuid
            duration:
              type: number
              default: 5
            aspectRatio:
              type: string
              default: "16:9"
            resolution:
              type: string
              default: "720p"
            pollIntervalSeconds:
              type: integer
              default: 5
            maxPolls:
              type: integer
              default: 120
            maxShots:
              type: integer
              minimum: 1
              maximum: 3
              default: 3
            skipCompose:
              type: boolean
              default: false
            generateAssets:
              type: boolean
              default: true
            generateDerivedAssets:
              type: boolean
              default: false
        idempotencyKey:
          type: string
          maxLength: 200
    CancelWorkflowRunRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional human-readable cancellation reason.
...